processor/requirements.txt
fastapi
uvicorn
pandas
mysql-connector-python
jinja2